+++ haudiotagger/lib/src/rust/api/tag.dart
@@ -7,6 +7,7 @@ import '../frb_generated.dart';
 import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 import 'picture.dart';
 
+// These functions are ignored because they are not marked as `pub`: `is_empty`
 // These function are ignored because they are on traits that is not defined in current crate (put an empty `#[frb]` on it to unignore): `clone`, `eq`, `fmt`, `from`
 
 /// Represents the metadata of the file.
@@ -92,11 +93,6 @@ class Tag {
   });
 
   static Future<Tag> default_() => RustLib.instance.api.crateApiTagTagDefault();
-
-  /// Returns `true` if the tag has no data.
-  Future<bool> isEmpty() => RustLib.instance.api.crateApiTagTagIsEmpty(
-        that: this,
-      );
 
   @override
   int get hashCode =>